% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{celsify_temp}
\alias{celsify_temp}
\title{Modifies temp to Celcius if country is US}
\usage{
celsify_temp(dat)
}
\arguments{
\item{dat}{A data frame with countries and temperatures}
}
\value{
A data frame where the temperatures is in Farenheit
}
\description{
Modifies temp to Celcius if country is US
}
