% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSampling.R
\name{CheckSampling}
\alias{CheckSampling}
\title{Check Sampling type function}
\usage{
CheckSampling(fixedrows, gstandard)
}
\arguments{
\item{fixedrows}{Boolean. Indicate if sample rows are fixed beforehand. Default is TRUE.}

\item{gstandard}{Text. Indicate if there are a Gold Standard by Rows or columns. Only first letter matter without Case sensitivity. Options are: "N" for None, "R" for in Rows and "C" for in Columns. Default is "N".}
}
\description{
This function checks that fixedrows and gstandard parameters are correct. Also return a value indicating if it is or not valid and what kind of output should be shown at the end of the execution.
}
\examples{
CheckSampling(TRUE,"rows")
CheckSampling(TRUE,"Columns")
}
\keyword{check}
\keyword{fixedrows}
\keyword{gstandard}
