% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohen_Kappa.R
\name{Cohen_Kappa}
\alias{Cohen_Kappa}
\title{Calculate Cohen's Kappa coefficient function}
\usage{
Cohen_Kappa(mx)
}
\arguments{
\item{mx}{Matrix. Agreement contingency table to perform calculations}
}
\description{
This function perform Cohen's Kappa coefficient calculations. The function provide the Kappa coefficient and SE with the provided.
}
\examples{
Cohen_Kappa(matrix(c(50,10,10,20),2,2))
}
\keyword{Cohen}
\keyword{Kappa}
\keyword{mx}
