\name{generate_random_random_email}
\alias{generate_random_random_email}
\title{Generate Completely Random Email Address}
\author{Prashant Kumar Jha (\email{jha.prashant1106@gmail.com})}
\description{
  The \code{generate_random_random_email} function generates a completely random email address.
}
\usage{
generate_random_random_email()
}
\details{
  This function generates a completely random email address with random characters for the local-part and domain.
}
\arguments{None}
\seealso{
\code{\link{generate_random_dataframe}},
\code{\link{generate_random_vector}}, \code{\link{generate_random_ip}},
\code{\link{generate_random_mobile_number}},
\code{\link{generate_random_email}}
}
\value{
  A character string representing a completely random email address.
}
\examples{
# Example usage of generate_random_random_email function

generate_random_random_email()
}