% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAsinDeltaParams.R
\name{GetAsinDeltaParams}
\alias{GetAsinDeltaParams}
\alias{print.GetAsinDeltaParams}
\title{Calculate Asintotic Delta related parameters function}
\usage{
GetAsinDeltaParams(mx, fixedrows = TRUE)

\method{print}{GetAsinDeltaParams}(x, ...)
}
\arguments{
\item{mx}{Matrix. Agreement contingency table to perform calculations}

\item{fixedrows}{Boolean. Indicate if sample rows are fixed beforehand.}

\item{x}{List produced by GetAsinDeltaParams}

\item{...}{Other print options}
}
\value{
\code{NULL}
}
\description{
This function perform all needed calculations to get all Delta related parameters, for a 2x2 matrix. All calculations are asintotics.
}
\examples{
GetAsinDeltaParams(matrix(c(60,10,10,20),2,2),TRUE)
}
\keyword{Asintotic}
\keyword{Delta}
\keyword{fixedrows}
\keyword{mx}
