% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCovariance.R
\name{GetCovariance}
\alias{GetCovariance}
\title{Calculate Covariance function}
\usage{
GetCovariance(mx, Delta, Pi, B, k)
}
\arguments{
\item{mx}{Matrix. Modified matrix to have a solution Usually GetMx$M2.}

\item{Delta}{Vector. Each element indicate the probability of recognize an element i.}

\item{Pi}{Vector. Each element indicate the probability of classify at random an element in category i.}

\item{B}{Double. Numerical solution to the equation given by the model.}

\item{k}{Integer. Dimension of the problem.}
}
\description{
This function calculate covariance for combinations Cov(Delta,Delta), Cov(Delta,Pi) and Cov(Pi,Pi).
}
\examples{
GetCovariance(mx = matrix(c(1,0,0,0,2,0,0,0,3),3,3), Delta = c(1,1,1), Pi = NULL,B = 0, k = 3)
GetCovariance(mx = matrix(c(1.5,2.5,0.5,3.5,4.5,0.5,0.5,0.5,1.5),3,3), Delta = c(-0.2662395,  0.2047577,  0.5664672), Pi = c(0.42564365, 0.49700867, 0.07734769), B =  14.01247, k = 3)
GetCovarianze(mx = matrix(c(60,0,3,2,50,1,3,2,79),3,3), Delta = c( 0.8945724, 0.9522836, 0.8962094), Pi = c( 0.2703707, 0.1939561, 0.5356732), B = 17.94867, k = 3)
}
\keyword{B}
\keyword{Covarianze}
\keyword{Delta}
\keyword{Pi}
\keyword{k}
\keyword{mx}
