% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDeltaParams.R
\name{GetDeltaParams}
\alias{GetDeltaParams}
\title{Calculate Delta related parameters function}
\usage{
GetDeltaParams(mx, fixedrows, Delta, Pi, B, k, Cov)
}
\arguments{
\item{mx}{Matrix. Agreement contingency table to perform calculations}

\item{fixedrows}{Boolean. Indicate if sample rows are fixed beforehand.}

\item{Delta}{Vector. Each element indicate the probability of recognize an element i.}

\item{Pi}{Vector. Each element indicate the probability of classify at random an element in category i.}

\item{B}{Double. Numerical solution to the equation given by the model.}

\item{k}{Integer. Dimension of the problem.}

\item{Cov}{Matrix. Covarianze matrix of Delta.}
}
\description{
This function perform all needed calculations to get all Delta related parameters. For do the exact calculations some variables previously calculated are needed.
}
\examples{
GetDeltaParams(...)
}
\keyword{B}
\keyword{Cov}
\keyword{Delta}
\keyword{Pi}
\keyword{fixedrows}
\keyword{k}
\keyword{mx}
