% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDeltaProblemType.r
\name{GetDeltaProblemType}
\alias{GetDeltaProblemType}
\title{Get Delta problem type function}
\usage{
GetDeltaProblemType(Mx)
}
\arguments{
\item{Mx}{Matrix. Matrix reduced.}
}
\description{
This function apply Test to identify where are the solution located. We have mainly 3 situations for k > 2 and 2 for k = 2. For k > 2 we have:
DN2 = No estimators in the boundary. 
DN1 = Some estimator in the boundary and global agreement imaginary. 
DN0 = Any other case 
For k = 2 we have:
DA0 = Some estimator in the boundary. 
DA1 = Any other case
}
\examples{
GetDeltaProblemType(matrix(c(1,2,0,3,4,0,0,0,1),3,3))
GetDeltaProblemType(matrix(c(1,0,0,0,2,0,0,0,3),3,3))
}
\keyword{Mx}
