% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGoodness.r, R/print.GetGoodness.R
\name{GetGoodness}
\alias{GetGoodness}
\alias{print.GetGoodness}
\title{Calculate Goodness of fit function}
\usage{
GetGoodness(mx, Pi, Delta)

\method{print}{GetGoodness}(x, ...)
}
\arguments{
\item{mx}{Matrix. Modified matrix to have a solution. Usually GetMx$M1 for k>2 and GetMx$M2 in case of k = 2.}

\item{Pi}{Vector. Each element indicate the probability of classify at random an element in category i.}

\item{Delta}{Vector. Each element indicate the probability of recognize an element i.}

\item{x}{List produced by GetGoodness}

\item{...}{Other print options}
}
\value{
\code{NULL}
}
\description{
This function provide an Chi-square test for the given matrix, Delta and Pi provided.
}
\examples{
GetGoodness(mx = matrix(c(1,0,0,0,2,0,0,0,3),3,3), Delta = c(1,1,1), Pi = NULL)
GetGoodness(mx = matrix(c(1.5,2.5,0.5,3.5,4.5,0.5,0.5,0.5,1.5),3,3), 
            Delta = c(-0.2662395,  0.2047577,  0.5664672), 
            Pi = c(0.42564365, 0.49700867, 0.07734769))
GetGoodness(mx = matrix(c(60,0,3,2,50,1,3,2,79),3,3), 
            Delta = c( 0.8945724, 0.9522836, 0.8962094), 
            Pi = c( 0.2703707, 0.1939561, 0.5356732))
}
\keyword{Delta}
\keyword{Goodness}
\keyword{Pi}
\keyword{mx}
