% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKappaProblemType.r
\name{GetKappaProblemType}
\alias{GetKappaProblemType}
\title{Get Kappa problem type function}
\usage{
GetKappaProblemType(Mx)
}
\arguments{
\item{Mx}{Matrix. Matrix reduced.}
}
\description{
This function apply Test to identify where kappa solutions are placed
K0 = Full agreement (diagonal matrix)
K1 = Any other case
}
\examples{
GetKappaProblemType(matrix(c(1,2,0,3,4,0,0,0,1),3,3))
GetKappaProblemType(matrix(c(1,0,0,0,2,0,0,0,3),3,3))
}
\keyword{Mx}
