% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetM1.R
\name{GetM1}
\alias{GetM1}
\title{Get reduced matrix (M1) function}
\usage{
GetM1(datatable)
}
\arguments{
\item{datatable}{Matrix. Expected to be square matrix with at least 2 rows (columns), non negative values and at least an element different of zero.}
}
\description{
This function reduce matrix provided by the user deleting missing categories, those j where \sum_{i=1}^n a_{ji} = \sum_{i=1}^n a_ij = 0.
Also provide a list of the categories deleted and provides the new size of the problem
}
\examples{
GetM1(matrix(c(1,2,0,3,4,0,0,0,0),3,3))
}
\keyword{M1}
\keyword{datatable}
