% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{localize_beach}
\alias{localize_beach}
\title{Find whether the beach is in US or UK}
\usage{
localize_beach(dat)
}
\arguments{
\item{dat}{A data frame listing beaches}
}
\value{
A data frame with a column indicating US or UK
}
\description{
Find whether the beach is in US or UK
}
