\name{LOG}
\alias{LOG}
\alias{log}
\title{
LOG
}
\description{
The logs of Diagnose and Detect Markers in Extracellular Circulating \cr \cr
logs(); \cr \cr
}
\details{
logs();
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \email{yujoan_2001@163.com} \cr 
Huansheng Cao (UGA) \email{hshcao@uga.edu} \cr 
Guoqing Liu (IMUST & UGA) \email{gqliu@uga.edu} \email{gqliu1010@163.com} \cr 
Xiufeng Liu (GZUCM & UGA) \email{xfliu@uga.edu} \email{liu_xf@gzucm.edu.cn} \cr 
Hao Wu (BIT & UGA) \email{wuhao@uga.edu} \email{wuhao@bit.edu.cn} \cr 
Yan Wang (JLU & UGA) \email{wy6868@hotmail.com} \cr 
Ying Xu (JLU & UGA) \email{xyn@uga.edu} \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
citation("delta");
}
\keyword{ LOG }
\seealso{
\code{\link{DELTA-package}} \code{\link{DELTA-method}} \code{\link{getBC-method}} \cr
}
\examples{

logs();

# LOGs: 
# DELTA 0.0, 04/28/2013, Project begin;
# DELTA 0.1, 12/21/2014, the system is send out for testing;
# DELTA 0.8, 08/04/2016, DELTA's C-version and R-version are published.

}
