% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{outfile_path}
\alias{outfile_path}
\title{Adds a time stamp on a file}
\usage{
outfile_path(infile)
}
\arguments{
\item{infile}{A csv file}
}
\value{
A csv file with time stamp in title
}
\description{
Adds a time stamp on a file
}
