% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDeltaParams.R
\name{GetDeltaParams}
\alias{GetDeltaParams}
\title{Calculate Delta related parameters function}
\usage{
GetDeltaParams(mx, Delta, Pi, k)
}
\arguments{
\item{mx}{Matrix. Agreement contingency table to perform calculations}

\item{Delta}{Vector. Each element indicate the probability of recognize an element i.}

\item{Pi}{Vector. Each element indicate the probability of classify at random an element in category i.}

\item{k}{Integer. Dimension of the problem.}
}
\description{
This function perform all needed calculations to get all Delta related parameters. For do the exact calculations some variables previously calculated are needed.
}
\examples{
GetDeltaParams(mx = matrix(c(60,0,3,2,50,1,3,2,79),3,3), 
			Delta = c( 0.8945724, 0.9522836, 0.8962094), 
              Pi = c( 0.2703707, 0.1939561, 0.5356732), k = 3)
}
\keyword{B}
\keyword{Cov}
\keyword{Delta}
\keyword{Pi}
\keyword{fixedrows}
\keyword{k}
\keyword{mx}
