% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKappa.R
\name{GetKappa}
\alias{GetKappa}
\alias{print.GetKappa}
\title{Calculate Cohen's Kappa coefficient function}
\usage{
GetKappa(mx)

\method{print}{GetKappa}(x, ...)
}
\arguments{
\item{mx}{Matrix. Agreement contingency table to perform calculations}

\item{x}{List produced by GetKappa}

\item{...}{Other print options}
}
\value{
\code{NULL}
}
\description{
This function perform Cohen's Kappa coefficient calculations. The function provide the Kappa coefficient and SE.
}
\examples{
GetKappa(matrix(c(50,10,10,20),2,2))
}
\keyword{Cohen}
\keyword{Kappa}
\keyword{mx}
