% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMx.R
\name{GetMx}
\alias{GetMx}
\title{Get matrix of the problem (Mx) function}
\usage{
GetMx(M1)
}
\arguments{
\item{M1}{Matrix. Initial matrix without missing categories.}
}
\description{
This function produce 3 new auxiliar matrix. Those matrix are always defined, but depending of the problem they will be used or not.
All the auxiliar matrices are created to be able to solve the problem and avoid issues with solutions in the boundary or not completly defined.
}
\details{
The matrix are defined as follows
 - M2: Extended M1 with $c_33$ of 1 and increased by 0.5
 - M3: M1 increased by 0.5
 - M4: M1 increased by 1

In case of M1 is not 2x2, M2 and M3 are the same matrix.
}
\examples{
GetMx(matrix(c(1,2,0,3,4,0,0,0,1),3,3))
}
\keyword{M1}
\keyword{Mx}
